Create EKS terraform module
This standardize EKS cluster creation and add several standard packages using
helm. The module utilizes the terraform-modules/terraform-aws-eks
upstream modules but it wraps the module by defaulting several of the
options as well as installing a couple of helm charts.

Once deployed the cluster will install several helm packages that we
usually need to have installed on every EKS cluster.