Creating velero-bucket module
This module will be a replacement for mozilla-it/tf-ark-backups, this
module will create an S3 bucket and an IAM role that will grant
permission to write to the bucket.

The IAM role utilizes the new IAM Role for Service Accounts feature that
was introduced last year which allows us to annotate a Service Account
to grant IAM access to the S3 bucket.

See [blog
post](https://aws.amazon.com/blogs/opensource/introducing-fine-grained-iam-roles-service-accounts/)
for more details