Fixes an issue where helm CRD gets installed over and over again
There seems to be an issue with the flux-helm-operator chart that causes
helm to attempt to install the CRD with every run. Based on what my
understanding of the chart the `createCRD` option is meant for helm v2
but will also work with helm v3 if you give helm the `--skip-crds` flag

I made a note to figure out if this will bite us in the future and also
made a note to try to figure out a method of installing he CRD not via
helm. Also sneaked in an additional flux option to the module