Split out flux helm installation to its own module
To help facilitate support for both AWS and GCP the installation of flux
and flux helm operator has been moved to its own folder as a module so
that its possible for anyone with a kubernetes cluster who wants to use
this can use it by including the module.

README has steps on how to set this up.