Adding an optional kubernetes permission object
This will be readonly/view.
It was easiest to add here because the kube perms don't correctly get exposed by the eks module.
Choose to keep this at the same level as other features to keep the code simple.
This assumes we won't have that many rbac objects to create, it'll be a pain to manage if we get to 10 for example.=